Applying code review comments
Possible fix to CI failures

Moving Segment class out of ConcurrentQueue
